elf: Move tests to separate file
Move the tests module to a separate file.
